backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
supabase
python-dotenv
httpx
